@@ -5,36 +5,37 @@
   "dependencies": {
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.57",
-    "@mui/icons-material": "^5.14.18",
-    "@mui/material": "^5.14.18",
+    "@mui/icons-material": "^5.15.15",
+    "@mui/lab": "^5.0.0-alpha.170",
+    "@mui/material": "^5.15.15",
+    "@mui/system": "^5.15.15",
+    "@react-keycloak/web": "^3.4.0",
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.6.0",
     "axios": "^0.21.1",
     "bootstrap": "^5.3.2",
+    "caniuse-lite": "^1.0.30001669",
+    "globals": "^15.1.0",
     "install": "^0.13.0",
+    "keycloak-js": "^26.0.1",
     "luxon": "^1.25.0",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.9.0-beta.1",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-error-boundary": "^3.1.0",
     "react-quill": "^2.0.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "recoil": "^0.1.2",
+    "recoil": "^0.1.3",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "vite",
+    "build": "tsc && vite build",
     "lint": "run-p lint:*",
     "lint:eslint": "eslint --fix ./src",
-    "lint:prettier": "prettier --write ./src"
+    "lint:prettier": "prettier --write ./src",
+    "serve": "vite preview"
   },
   "eslintConfig": {
     "extends": [
@@ -55,10 +56,15 @@
     ]
   },
   "devDependencies": {
+    "@types/react-router-dom": "^5.3.3",
+    "@vitejs/plugin-react": "^4.2.1",
+    "eslint": "^9.1.1",
+    "eslint-plugin-react": "^7.34.1",
     "husky": "^4.3.6",
     "lint-staged": "^10.5.3",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.2.1"
+    "prettier": "^2.2.1",
+    "vite": "^5.2.10"
   },
   "husky": {
     "hooks": {
@@ -70,6 +76,6 @@
     "*.{js,css,md}": "prettier --write"
   },
   "engines": {
-    "node": "14.15.x"
+    "node": "^20.16.x"
   }
 }
