@@ -8,19 +8,17 @@ import Checkbox from '@mui/material/Checkbox';
 import Button from '@mui/material/Button';
 import Paper from '@mui/material/Paper';
 import PropTypes from 'prop-types';
-import { useState } from 'react';
-import { useEffect } from 'react';
+import { useState, useEffect  } from 'react';
+
 import { getGrupoByCursoId } from '../../services/Grupo';
 import { Container } from '@mui/system';
-import Box from '@mui/material/Box';
-import Typography from '@mui/material/Typography';
-import Modal from 'react-bootstrap/Modal';
-import { Divider } from '@material-ui/core';
+import Divider from '@mui/material/Divider';
 
+// Proposito:  Devuelve los elementos en a que no están en b.
 function not(a, b) {
   return a.filter((value) => b.indexOf(value) === -1);
 }
-
+// Proposito: Devuelve los elementos que están en ambos, a y b
 function intersection(a, b) {
   return a.filter((value) => b.indexOf(value) !== -1);
 }
@@ -29,12 +27,11 @@ export default function ListaDeGrupos({
   idCurso,
   show,
   closeModal,
-  setGruposParaTrabajo,
+  setGruposParaTrabajo, //Función para establecer los grupos seleccionados para el trabajo práctico.
 }) {
-  const [checked, setChecked] = React.useState([]);
-
-  const [gruposParaTp, setGruposParaTp] = useState([]);
-  const [gruposDeCurso, setGruposDeCurso] = useState([]);
+  const [checked, setChecked] = React.useState([]); //checked: Lista de grupos seleccionados (checkeados).
+  const [gruposParaTp, setGruposParaTp] = useState([]); //gruposParaTp: Grupos seleccionados para el trabajo práctico.
+  const [gruposDeCurso, setGruposDeCurso] = useState([]);//: Grupos disponibles en el curso.
 
   const gruposParaTpChecked = intersection(checked, gruposParaTp);
   const gruposDeCursoChecked = intersection(checked, gruposDeCurso);
@@ -46,14 +43,10 @@ export default function ListaDeGrupos({
 
   useEffect(() => {
     async function fetchGrupos() {
-      console.log('show.', show);
       setGruposDeCurso([]);
-
       const getFunction = !show ? [] : getGrupoByCursoId(idCurso);
-
       try {
         const grupos = await getFunction;
-
         setGruposDeCurso(grupos);
       } catch (err) {
         console.log('Ocurrio este error.', err);
@@ -124,7 +117,7 @@ export default function ListaDeGrupos({
             <ListItem
               key={value._id}
               role="listitem"
-              button
+              
               onClick={handleToggle(value)}
             >
               <ListItemIcon>
@@ -160,7 +153,7 @@ export default function ListaDeGrupos({
   return (
     <>
       <Grid container spacing={2} justifyContent="center" alignItems="center">
-        <Grid item>{customList(gruposParaTp, 'Grupos para crear TP')}</Grid>
+        <Grid item>{customList(gruposParaTp, 'Grupos para asignar TP')}</Grid>
         <Grid item>
           <Grid container direction="column" alignItems="center">
             <Button
@@ -212,10 +205,18 @@ export default function ListaDeGrupos({
           display: 'flex',
           flexDirection: ' row',
           justifyContent: 'space-evenly',
-          paddingTop: '4%',
+          paddingTop: '2%',
+          paddingBottom: '2%', 
         }}
       >
-        <Button onClick={agregarGrupo} variant="contained" color="success">
+        <Button onClick={agregarGrupo} 
+        variant="contained" 
+        color="success" 
+          sx={{ 
+            backgroundColor: '#c5e1a5', 
+            color: '#000000', 
+            '&:hover': { backgroundColor: '#b0d38a' 
+            } }}>
           Agregar Grupos a Tp
         </Button>
       </Container>
@@ -226,6 +227,6 @@ export default function ListaDeGrupos({
 ListaDeGrupos.propTypes = {
   idCurso: PropTypes.string.isRequired,
   show: PropTypes.bool.isRequired,
-  closeModal: PropTypes.element.isRequired,
-  setGruposParaTrabajo: PropTypes.element.isRequired, // Ajusta el tipo y la obligatoriedad según tu lógica de uso
+  closeModal: PropTypes.func.isRequired,// Cambia el tipo a func
+  setGruposParaTrabajo: PropTypes.func.isRequired, // Ajusta el tipo y la obligatoriedad según tu lógica de uso
 };
